Fix FileChooser portal to be able be opened using show() method
Summary:
My assumption before was that show() method calls exec(), but it seems to be vice-versa
and so QML applications using QFileDialog were not able to open dialogs as they use
visible property, which in turn calls show() method. I made the show() method to call
new openPortal() method where I moved the actuall DBus call from exec() method. The exec()
method now internally calls show(), at least this is my assumption and it seems to behave
like that from my testing and on top of that it just adds a QEventLoop so the call is
blocking as it was before.

Reviewers: apol

Reviewed By: apol

Subscribers: ngraham

Differential Revision: https://phabricator.kde.org/D9842