Use the new freedesktop-post.sh script
Summary:
This script cleans up a bunch of things in the manifest so that we
don't have to copy it everywhere (and update everywhere).

The script in the current base also has a tweak to the fc-cache call
to work better with the new fontconfig cache format. Additionally we no
longer cleanup /cache as that also broke the font cache.

Reviewers: apol

Differential Revision: https://phabricator.kde.org/D9915