Use QGnomePlatform as default platform theme on GNOME
Qt6 now use QDirIterator instead of QDir when loading plugins and
because of that, it is no longer guaranteed that the plugins will
load in an alphabetical order, which is what we relied on that
QGnomePlatform will be loaded before QGtk3. For that reason we
need to attempt to load QGnomePlatform directly and have QGtk3 as
another fallback.