Add ocean-sound-theme
I'm the maintainer of the Ocean sound theme, and it came to my attention that KDE Flatpak apps cannot play sounds from the theme because it's not included in the Flatpak KDE Runtime.

Even though I'm the maintainer of the sound theme, I'm not a developer, so I'm unsure if this is the right approach. Any help would be appreciated.

Looking through the file I noticed all things for Plasma are currently in version 6.3.4, so in the commit section I put `6f564989422976f51346d8273b827bf647adec52`, which points to the [commit made for version 6.3.4](https://invent.kde.org/plasma/ocean-sound-theme/-/commit/6f564989422976f51346d8273b827bf647adec52). I based this patch on the description for other Plasma components that were in the same file, but I'm unsure if anything else is necessary. I don't know how to test if this works and couldn't find any instructions, so please review carefully and let me know if there's anything I need to change.

Also, do I need to make another merge request for the 6.8 branch?

Thanks! @aacid 

BUG:409300